+++ packages/koot/React/inject/styles.js
@@ -29,17 +29,6 @@ module.exports = ({
 }) => {
     if (typeof injectCache[styles] === 'undefined') {
         injectCache[styles] = getExtracted(localeId, compilation);
-
-        if (process.env.WEBPACK_BUILD_ENV === 'dev') {
-            injectCache[styles] +=
-                `<style id="__koot-react-hot-loader-error-overlay">` +
-                `.react-hot-loader-error-overlay div {` +
-                `    z-index: 99999;` +
-                `    font-size: 14px;` +
-                `    line-height: 1.5em;` +
-                `}` +
-                `</style>`;
-        }
     }
 
     // if (needInjectCritical && typeof filemap['critical.css'] === 'string') {